# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanIes",
        "contact": {
            "email": "derex@outlook.com.br"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/v1/release": {
            "get": {
                "tags": [
                    "Release"
                ],
                "summary": "get all release",
                "description": "get all release",
                "operationId": "30c75eaa4ad8a0caa07b244e088073b3",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Release"
                ],
                "summary": "register a new release",
                "description": "register a new release",
                "operationId": "565a83d739aefa4393fcf81a2304ba95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "value": {
                                        "description": "value",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "date": {
                                        "description": "date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "voucher": {
                                        "description": "The data block for encryption/decryption",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "status",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/release/{release}/attach": {
            "post": {
                "tags": [
                    "Release"
                ],
                "summary": "associate permission to release",
                "description": "associate permission to release",
                "operationId": "1e50237f44d286cc544d962a8d70fa06",
                "parameters": [
                    {
                        "name": "release",
                        "in": "path",
                        "description": "release identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permission_id": {
                                        "description": "permission id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/release/{release}": {
            "put": {
                "tags": [
                    "Release"
                ],
                "summary": "update a release by id",
                "description": "update a release by id",
                "operationId": "updateRelease",
                "parameters": [
                    {
                        "name": "release",
                        "in": "path",
                        "description": "release identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "value": {
                                        "description": "value",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "date": {
                                        "description": "date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "voucher": {
                                        "description": "The data block for encryption/decryption",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "status",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Release"
                ],
                "summary": "delete a release by id",
                "description": "delete a release by id",
                "operationId": "deleteRelease",
                "parameters": [
                    {
                        "name": "release",
                        "in": "path",
                        "description": "release identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    }
}